feat(store): impl token

+++ src/storage/src/types.rs
@@ -1,9 +1,9 @@
 pub type NodeId = u64;
 pub type RelationshipId = u64;
 
-pub type LabelId = u32;
-pub type RelationshipTypeId = u32;
-pub type PropertyKeyId = u32;
+pub type LabelId = u16;
+pub type RelationshipTypeId = u16;
+pub type PropertyKeyId = u16;
 
 pub type PropertyKey = String;
 pub type Label = String;
